+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All changes to this project are documented in this file. This project adheres to
 ### Changed
 - Update eslint-plugin-react dependency to ^5.2.2
 - Fix missing `jsx-space-before-closing` rule
+- Improve the browser, ES6, Node, and React configuration docs
 
 ## [7.4.0] - 2016-07-04
 ### Added