+++ CHANGELOG.md
@@ -18,6 +18,7 @@ All changes to this project are documented in this file. This project adheres to
 - Update eslint-plugin-react dependency to ^5.2.2
 - Fix missing `jsx-space-before-closing` rule
 - Improve the browser, ES6, Node, and React configuration docs
+- Improve the parser options configuration docs
 
 ## [7.4.0] - 2016-07-04
 ### Added